HTML Code to demonstrate Canvas Element
This code creates a canvas and then uses JavaScript to draw a circle on the canvas. The circle’s position is updated every second, causing it to move. If the circle hits the edge of the canvas, it changes direction. The color of the circle is also changed every second, creating a fun and colorful animation.